moved ReadPlan, WritePlan and SetRootDir out of internals. Made SetRootDir into a method.
