+++ types/plan/testdata/setrootdir/nodejsplan.yaml
